Testing Actor Replies
The ExpectMsg feature of the TestKit allows us to assert that we got back a message, and also that the message is of the correct type.

When we use ExpectMsg, we get a default timeout of 3 seconds, but if we need to, we can override this with a custom timeout.